refactor: add scanner, rewrite grammar and queries
For correct matching the `type_value_definition` token I've written the
scanner which can lookup more further than can be declarable in grammar
witout including anchor tokens (in this case - opening parentheis).

With it the grammar was rewritten by removing pattern matching for
`type_value_identifier` (and use the scanner for this) and added
`constructor` token. With this token the indentation will be always
correct.

Then, the queries was rewritten by reducing trash code with small but
very powerful matching. The highlights, indents and folds works
perfectly.